Move `assets/js/components/surveys/CurrentSurvey.js` to its own module.
