+++ assets/js/components/surveys/CurrentSurvey/index.test.js
@@ -19,16 +19,16 @@
 /**
  * Internal dependencies
  */
-import { CORE_USER } from '../../googlesitekit/datastore/user/constants';
-import { CORE_FORMS } from '../../googlesitekit/datastore/forms/constants';
+import { CORE_USER } from '../../../googlesitekit/datastore/user/constants';
+import { CORE_FORMS } from '../../../googlesitekit/datastore/forms/constants';
 import {
 	render,
 	fireEvent,
 	createTestRegistry,
 	waitFor,
 	provideCurrentSurvey,
-} from '../../../../tests/js/test-utils';
-import CurrentSurvey from './CurrentSurvey';
+} from '../../../../../tests/js/test-utils';
+import CurrentSurvey from '.';
 import {
 	singleQuestionMultiSelect,
 	singleQuestionOpenText,
@@ -37,7 +37,7 @@ import {
 	multiQuestionSurvey,
 	multiQuestionConditionalSurvey,
 	invalidQuestionTypeSurvey,
-} from './__fixtures__';
+} from '../__fixtures__';
 
 // Text input should only allow up to 200 characters of input.
 const STRING_100_CHARACTERS =
